refactor: rename partition_mapper package as partition_mappers
