+++ airflow-core/src/airflow/partition_mappers/__init__.py
