+++ airflow-core/src/airflow/partition_mappers/base.py
