+++ airflow-core/src/airflow/serialization/helpers.py
@@ -27,7 +27,7 @@ from airflow.configuration import conf
 from airflow.settings import json
 
 if TYPE_CHECKING:
-    from airflow.partition_mapper.base import PartitionMapper
+    from airflow.partition_mappers.base import PartitionMapper
     from airflow.timetables.base import Timetable as CoreTimetable
 
 
@@ -162,4 +162,4 @@ class PartitionMapperNotFound(ValueError):
 
 def is_core_partition_mapper_import_path(importable_string: str) -> bool:
     """Whether an importable string points to a core partition mapper class."""
-    return importable_string.startswith("airflow.partition_mapper.")
+    return importable_string.startswith("airflow.partition_mappers.")
