+++ airflow-core/tests/unit/serialization/test_serialized_objects.py
@@ -733,13 +733,13 @@ def test_encode_partition_mapper():
 
     partition_mapper = IdentityMapper()
     assert encode_partition_mapper(partition_mapper) == {
-        Encoding.TYPE: "airflow.partition_mapper.identity.IdentityMapper",
+        Encoding.TYPE: "airflow.partition_mappers.identity.IdentityMapper",
         Encoding.VAR: {},
     }
 
 
 def test_decode_partition_mapper():
-    from airflow.partition_mapper.identity import IdentityMapper as CoreIdentityMapper
+    from airflow.partition_mappers.identity import IdentityMapper as CoreIdentityMapper
     from airflow.sdk import IdentityMapper
     from airflow.serialization.decoders import decode_partition_mapper
     from airflow.serialization.encoders import encode_partition_mapper
