+++ task-sdk/src/airflow/sdk/__init__.py
@@ -108,8 +108,8 @@ if TYPE_CHECKING:
     from airflow.sdk.definitions.decorators.task_group import task_group
     from airflow.sdk.definitions.edges import EdgeModifier, Label
     from airflow.sdk.definitions.param import Param, ParamsDict
-    from airflow.sdk.definitions.partition_mapper.base import PartitionMapper
-    from airflow.sdk.definitions.partition_mapper.identity import IdentityMapper
+    from airflow.sdk.definitions.partition_mappers.base import PartitionMapper
+    from airflow.sdk.definitions.partition_mappers.identity import IdentityMapper
     from airflow.sdk.definitions.taskgroup import TaskGroup
     from airflow.sdk.definitions.template import literal
     from airflow.sdk.definitions.timetables.assets import (
@@ -162,7 +162,7 @@ __lazy_imports: dict[str, str] = {
     "DeltaTriggerTimetable": ".definitions.timetables.trigger",
     "EdgeModifier": ".definitions.edges",
     "EventsTimetable": ".definitions.timetables.events",
-    "IdentityMapper": ".definitions.partition_mapper.identity",
+    "IdentityMapper": ".definitions.partition_mappers.identity",
     "Label": ".definitions.edges",
     "Metadata": ".definitions.asset.metadata",
     "MultipleCronTriggerTimetable": ".definitions.timetables.trigger",
@@ -170,7 +170,7 @@ __lazy_imports: dict[str, str] = {
     "Param": ".definitions.param",
     "ParamsDict": ".definitions.param",
     "PartitionedAssetTimetable": ".definitions.timetables.assets",
-    "PartitionMapper": ".definitions.partition_mapper.base",
+    "PartitionMapper": ".definitions.partition_mappers.base",
     "PokeReturnValue": ".bases.sensor",
     "SecretCache": ".execution_time.cache",
     "SyncCallback": ".definitions.callback",
