+++ task-sdk/src/airflow/sdk/definitions/partition_mappers/base.py
